% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW_delay_distribution.R
\name{count_weekly_reporting}
\alias{count_weekly_reporting}
\title{DOE/DOR data aggregation}
\usage{
count_weekly_reporting(
  data,
  aggregation_date,
  max_week_delay,
  keep_weekly_prob = F
)
}
\arguments{
\item{data}{Raw data containing doe (Date of event), dor (Date of registration or reporting) and location_code. The columns must have these exact names.}

\item{aggregation_date}{Date of aggregation}

\item{max_week_delay}{Maximum weeks of delay taken into account. Must be greater than 2.}

\item{keep_weekly_prob}{Binary. If True, the final aggregated data has weekly reporting percentile.}
}
\value{
Aggregated weekly data with number of events happened each week and reporting delay distribution
}
\description{
Aggregates daily mortality reporting data to a weekly basis.
For more details see the help vignette:
}
\details{
\code{vignette("intro", package="nowcast")}
}
\examples{

data <- nowcast::data_fake_nowcasting_raw
aggregation_date <- as.Date("2020-01-01")
max_week_delay <- 4

weekly_counts <- count_weekly_reporting(data, aggregation_date, max_week_delay)

weekly_counts_p <- count_weekly_reporting(data, aggregation_date, max_week_delay, keep_weekly_prob = TRUE)
}
