% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NEW_delay_distribution.R
\docType{data}
\name{d_fake_weekly_doe_dor_aggregated}
\alias{d_fake_weekly_doe_dor_aggregated}
\title{Aggregate weekly data for events and delay distribution, expressed by weekly cumulatives.}
\format{
\describe{
\item{isoyearweek_event}{First date of every week}
\item{location_code}{Location code}
\item{monday_doe}{Date of the Monday for this isoyearweek}
\item{n_event}{Number of deaths this week, recorded until the aggregation date}
\item{n_cum_lagi}{Number of registered deaths within the next i weeks. 0 suggests the current week}
\item{p_cum_lagi}{Percentile of registrations within the next i weeks. 0 suggests the current week}
}
}
\usage{
d_fake_weekly_doe_dor_aggregated
}
\description{
Aggregate weekly data for events and delay distribution, expressed by weekly cumulatives.
}
\keyword{datasets}
