% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_death_county_data.R
\docType{data}
\name{data_fake_nowcasting_county_raw}
\alias{data_fake_nowcasting_county_raw}
\title{Fake data for mortalityregistration}
\format{
\describe{
\item{doe}{Date og event}
\item{dor}{Date of registration}
\item{location_code}{Location code}
}
}
\usage{
data_fake_nowcasting_county_raw
}
\description{
Fake data for mortalityregistration
}
\keyword{datasets}
