% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_fake_death_data.R
\docType{data}
\name{data_fake_nowcasting_raw}
\alias{data_fake_nowcasting_raw}
\title{Fake data for mortality registration}
\format{
\describe{
\item{doe}{Date og event}
\item{dor}{Date of registration}
}
}
\usage{
data_fake_nowcasting_raw
}
\description{
The fake data for mortality registration is generated by combining the daily
death counts and days of registration delay.
}
\details{
The expected death counts per day is 115 with standard deviation 10.

The expected registration delay is set to be 7 days.
}
\keyword{datasets}
