% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_mort.R
\name{nowcast_aggregate}
\alias{nowcast_aggregate}
\title{Aggregation of data for nowcasting}
\usage{
nowcast_aggregate(data, aggregation_date, n_week, pop_data = NULL)
}
\arguments{
\item{data}{Dataset containing doe (Date of event), dor (Date of registation) and location_code. The columns must have these exact names.}

\item{aggregation_date}{Date of aggregation}

\item{n_week}{Number of weeks to calculate the percentage of the total registraations. Must be larger og equal to 2 amd smaller than the total number of weeks in the dataset.}

\item{pop_data}{Population data, must contain a column called pop with the population data and a column with year and possibly week.}
}
\value{
Aggregated dataset with the percentiles of registered events within the last 52 weeks
}
\description{
Aggregates mortality data to a weekly basis.
Where the percentiles and number of deaths obtained after every week up to n_week is also given.
For more details see the help vignette:
}
\details{
\code{vignette("intro", package="nowcast")}
}
\examples{

data <- nowcast::data_fake_nowcasting_raw
data[doe < as.Date("2019-01-01")]
aggregation_date <- as.Date("2020-01-01")
n_week <- 6

data_aggregated <- nowcast_aggregate(data, aggregation_date, n_week)

}
