% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast_expand.R
\name{nowcast_correction_sim_neg_bin}
\alias{nowcast_correction_sim_neg_bin}
\title{Nowcast simmulation function when nowcast_correction_fn_negbin is used for correction.}
\usage{
nowcast_correction_sim_neg_bin(
  nowcast_correction_object,
  offset,
  n_sim = 500,
  date_0
)
}
\arguments{
\item{nowcast_correction_object}{object returned from function nowcast_correction_fn_expanded}

\item{offset}{Boolian value which is set to true if offset(log(pop)) is a part of the formula}

\item{n_sim}{Number of simulations}

\item{date_0}{Date of aggregation.}
}
\value{
simulations of the estimate made by the fitted models in nowcast_correction_fn
}
\description{
For more details see the help vignette.:
\code{vignette("nowcast", package="nowcast")}
}
\examples{
\dontrun{
data <- data.table::as.data.table(data_fake_nowcasting_county_aggregated)
n_week_adjusting <- 6
date_0 <- data[nrow(data),]$cut_doe #last date in the dataset, assume the dataset is ordered.
data <- data[cut_doe >= (date_0 - n_week_start*7 + 1), ]
nowcast_correction_object <- nowcast_correction_fn_negbin_mm(data, n_week_adjusting,
   offset = "log(pop)" , date_0 = date_0)
nowcast_sim <- nowcast_correction_sim_neg_bin(nowcast_correction_object,
    offset = "log(pop)", date_0 = date_0)
}
}
