% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nc_dataset}
\alias{print.nc_dataset}
\title{Print Method for nc_dataset Objects}
\usage{
\method{print}{nc_dataset}(x, ...)
}
\arguments{
\item{x}{An \code{nc_dataset} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for nc_dataset Objects
}
