% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nowcast_backtest}
\alias{print.nowcast_backtest}
\title{Print Method for Backtest Results}
\usage{
\method{print}{nowcast_backtest}(x, ...)
}
\arguments{
\item{x}{A \code{nowcast_backtest} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Backtest Results
}
