% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nowcast_result}
\alias{print.nowcast_result}
\title{Print Method for Nowcast Results}
\usage{
\method{print}{nowcast_result}(x, ...)
}
\arguments{
\item{x}{A \code{nowcast_result} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The input object, invisibly.
}
\description{
Print Method for Nowcast Results
}
